+++ src/features/legal/index.js
@@ -1,3 +1,4 @@
 export { default as LegalPageLayout } from './ui/LegalPageLayout';
+export { default as LegalReconsentModal } from './ui/LegalReconsentModal';
 export { default as PrivacyPolicyContent } from './ui/PrivacyPolicyContent';
 export { default as TermsOfServiceContent } from './ui/TermsOfServiceContent';